+++ packages/wallets/bitget-wallet/CHANGELOG.md
@@ -1,1 +1,18 @@
 # @web3-wallet/bitget-wallet
+
+## 1.0.0
+
+### Major Changes
+
+- @react-web3-wallet@v3
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-web3-wallet/core@3.0.0
+
+## 0.0.2
+
+### Patch Changes
+
+- publish bigget-wallet