+++ packages/wallets/trust-wallet/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @web3-wallet/trust-wallet
 
+## 3.0.0
+
+### Major Changes
+
+- @react-web3-wallet@v3
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-web3-wallet/core@3.0.0
+
 ## 2.4.1
 
 ### Patch Changes